tests/wpt/tests/svg/struct/UnknownElement/interface.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     version="1.1"
     width="100%" height="100%" viewBox="0 0 400 400">
<title>Interface of unknown element</title>
<h:link rel="help" href="https://svgwg.org/svg2-draft/struct.html#UnknownElement"/>
<h:script src="/resources/testharness.js"/>
<h:script src="/resources/testharnessreport.js"/>

<foobar id="target"/>

<h:script><![CDATA[
test(function() {
    var e = document.getElementById("target");
    assert_equals(Object.getPrototypeOf(e), SVGUnknownElement.prototype);
});
]]></h:script>
</svg>
